chore: Only run codecov on releases (#506)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Only run codecov upload on releases.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

Our code coverage is lagging behind because we rarely have a window to
upload the latest codecov on a free plan.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

The codecov CI is now running only on releases.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

Forces CI to run on tags only.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
- `codecov.yml`


## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Untested, will need to be tested upon next release.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [ ] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)