Improve NG dataset reproducibility; pin torchmetrics (#565)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->
These are small changes introduced in #546, split out from the work on
patch filtering.
> [!NOTE]  
> **tldr**: small changes to dataloader defaults and seeding, pin
`torchmetrics` version. <!-- Write a one sentence summary. -->

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->
1. Fix an issue with validation `num_workers` not being set from train
parameters when it is specified
2. Improve reproducibility for NG dataset - create a default random seed
and propagate it to transforms
3. Pin `torchmetrics` to avoid a parameter error (see #550).


## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
1. `data_model.py` and `ng_data_model.py`: Make a separate validator to
set the validation `num_workers` from train params, if needed
2. `ng_data_model.py`: Set default random seed, make a new validator to
propagate it to the transforms
3. `pyproject.toml`: Add `torchmetrics>=0.11.0,<1.5.0`

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->
I ran `bsd68_demo.ipynb` with NGDataset to check reproducibility (epoch
loss between runs) and parameter setting.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves #550 

---

**Please ensure your PR meets the following requirements:**

- [ ] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [ ] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>