Chore: add monkeypatching for MPS (#571)
This PR introduces a `monkeypatch` that ensures we are running tests
without `MPS` (macOS arm64 GPU). This is done by using pytest fixture
and changing pytorch variable.

The CI is back to being a single workflow for all three OS.

Note: I also fixed two pre-commit errors that came up, one of which
changes the signature of `configuration.model_dump` in order to avoid
issues everytime the `BaseModel` signature changes...