Feat: Generalizes `ZarrImageStack` (#594)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Zarr files can be arbitrarily organized, this PR makes the
`ZarrImageStack` more general and compatible with latest zarr-python.

This PR extends `ZarrImageStack` capabilities to arbitrary organization,
and adds a an `image_stack_loader` of sorts that follows a simple
heuristics:
- if we supply a path to a `.zarr` file, then either we instantiate all
arrays found in the file, or it is an OME-Zarr and we load only one
resolution level
- if we supply URIs
(`file://path/to/image.zarr/group/other_group/array`), then only the
arrays specified by the URIs are instantiated.

The scope of the PR is solely on the `ImageStack` mechanism, and will be
followed by other PRs:
- Compatibility with dataset: currently the default mechanism does not
allow for passing paths other that those ending with the correct file
extension.
- Prediction writer to Zarr

**Important notes and discussion**:
- One important change is that `zarr.open` is not called anymore in
`ZarrImageStack`, which contradicts the spririt of the NGDataset.
- The reason for this choice is that in order to collect groups and
arrays from a Zarr, we need to open it first... We could check how
expensive this operation is, and choose to open, list arrays and then
pass on the URIs for the `ZarrImageStack` to open again the file.
- The second reason is that regardless of the type of storage, new Zarr
docs basically encourage to use `zarr.open` which returns a `Group` (and
not a store)
- A subsequent PR could also add an `in_memory` keyword to the
`ZarrImageStack` for loading. But that would probably `ZarrImageStack`
to be in charge of opening the file, which may lead to loading of all
arrays. Alternatively, it could create a memory store and add the
existing array, but that would require copying data. I don't know if
there is a way to load in memory an array directly, without "accessing"
it.
- I have not found a solution for the moment to be able to explore all
arrays in a remote file (in the tests, `group.keys()` and
`group.array_keys() hang).
- URI and OME-Zarr validation are using a very minimal set of
constraints, which may not hold at all time.
- `create_zarr_image_stacks` could be refactored further, and
`zarr_utils` placed somewhere more meaningful


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

The current `ZarrImageStack` is only compatible with single-array
OME-Zarr with `"0"` resolution level. In the future, we want to be
flexible since Zarr comes in many flavors, and in particular support
having multiple arrays in a single zarr.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

Zarr image stack loader can instantiate multiple `ZarrImageStack` for a
single zarr, and allows specifying different arrays in various zarr
files as input.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- `zarr_utils` contains all the functions interacting with the Zarr
format

### Modified features or files

<!-- List important modified features or files. -->
- `ZarrImageStack` has been adapted to take a `Group` as input, and now
retains `chunk_size` and different `source` (URI-like) for future PRs

### Removed features or files

<!-- List removed features or files. -->
- all old OME functions.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Pytest


## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

All Zarr related code/notebook.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)