Feat: Enable Zarr CAREamics DataModule (#616)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Allow passing paths and URIs of Zarrs to the CAREamics
Lightning Datamodule.

Currently, the CAREamics Lightning Datamodule can receive arrays,
string, paths or list of the same types. When it comes to zarr, the
current design decision is to allow passing paths to groups inside a
zarr and collect all arrays that may be stored there. This is not
possible with the current path / file solver.

This PR brings the following changes:
- Refactor the array/path/file solving logic, move it to
`data_module_utils.py`, and add tests
- Add logic to solve string, path or URIs to Zarr groups and arrays
- Fix some bugs along the way that appeared when testing Zarr with the
data module


> [!NOTE]  
> Currently, here are potential pain points or blind angles, which we
could address in subsequent PR:
> - Matching names between input and target is based solely on sorting
the names of the arrays, not by checking that they agree
> - Online zarr and OME have not been tested from the DataModule level
(only at the `ZarrImageStack` level)
> - Passing a mixed list of path to zarr files and URIs will result in
only the type of the first element to be retained. I don't like that,
but I didn't want to make the function more complex for now. This should
be addressed in a subsequent PR (or now if that is the reviewer's
opinion).
> -The existence of the arrays or groups corresponding to the URIs is
not checked (so as to not open the zarr file)


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

In order to use URIs
(`file://data/folder/my_images.zarr/a_group/another_group`), which allow
a high level of granularity and flexibility in defining which arrays
should be used for training/validation/prediction, the logic of paths
solving in the `CAREamicsDataModule` needed to take into account this
new possibility.

More specifically, this PR aims at enabling the following behaviours:

```python
path_to_zarr = ["path/to/zarr/archive.zarr"] # load all arrays in the root
path_to_ome_zarr = ["path/to/zarr/ome_archive.zarr"] # load all arrays in the first level ("0" if not specified)
path_to_folder = ["path/to/folder/"] # load all arrays from the root of all zarr file in the folder
path_to_group = ["file://path/to/zarr/archive.zarr/group/subgroup"] # load all arrays from this group (obtained with g.store_path)
path_to_arrays = [
    "file://path/to/zarr/archive.zarr/group/array1",
    "file://path/to/zarr/archive.zarr/group/array2",
    "file://path/to/other_zarr/other_archive.zarr/other_group/array3",
] # load all specified arrays
```

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

- `CAREamicsDataModule` accepts `zarr` data type and offload solving the
paths/URIs to a new function in `data_module_utils.py`

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

As for the arrays, string and Path, the `zarr` datatype triggers a
series of checks:
- If `str` or `Path`, check if it corresponds to an existing `Path`,
then solve it using the normal path logic
- If `str` but not an existing `Path`, check if it is a URI, then solve
it
- If a list, repeat the same as above, but for a lists, based on the
first element of the list


## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- Moved all path solving feature to `data_module_utils.py`, which now
also contains Zarr logic, and refactor them a little

### Modified features or files

<!-- List important modified features or files. -->
- Changed the behaviour of the array collection in Zarr to limit itself
to the arrays in the group (rather than recursively go through the
sub-groups).

### Removed features or files

<!-- List removed features or files. -->
- Removed logic from the data module.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

- New tests have been implemented.
- BSD68 notebook tested for small iterations

- Resolves #598 

## Future improvements

- Currently one aspect is not great, but I decided not to make it more
complex for now: when inspecting a list of inputs, only the first
element is checked. So if the first element is a path to a zarr file, it
will only consider existing files. If it is a URI, it will only consider
URIs. There is no reason (except more complex code) to not just accept
all.
- Matching names of arrays / files for input/target solving.
- Checking existence of the arrays in the URI.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: Melisande Croft <63270704+melisande-c@users.noreply.github.com>