Refac(Next-Gen Dataset): Factory functions & Directory structure (#624)
## Description

> [!NOTE]  
> **tldr**: This PR reorganises the code, removing all the almost
duplicate functions we had between the dataset and patch extractor
factory functions. It also moves the `image_stack` directory into the
root `dataset_ng` directory, from inside of the `patch_extractor`
directory.

### Background - why do we need this PR?

Having the `image_stack` package inside the `patch_extractor` package
was a bit cumbersome, having it in the root means it is more accessible
and we can see all the components of the Next-Gen Dataset in the root.

All the factory functions that call each other was also confusing and
added too much unnecessary bloat.

### Overview - what changed?

This PR adds built-in implementations of the `ImageStackLoader` for
arrays, tiffs, iterated tiffs, zarrs, czi, and custom files. Now the
dataset creation pipeline is much simpler:
1. select the image stack loader (can also be custom)
2. select the patch extractor type (different for the iterable tiffs)
3. initialise the image stacks for the inputs, targets and masks
4. initialise the patch extractors from the image stacks for the inputs,
targests and masks
5. initialise the dataset from the patch extractors.

### Implementation - how did you implement the changes?

Mostly just moving code around. The image stack loader implementations
come from the `create_x_patch_extractor` functions. The zarr loader
comes from the implementation `create_zarr_image_stacks` function in
`image_stacks.zarr_utils`.


## Changes Made

### New features or files

- `select_image_stack_loader` function in `factory.py`
- `select_patch_extractor_type` function in `factory.py`
- `init_patch_extractor` function in `factory.py`

### Modified features or files

<!-- List important modified features or files. -->
- `create_dataset` function in `factory.py`
- moved `patch_extractor/image_stacks` -> `image_stacks`

### Removed features or files

<!-- List removed features or files. -->
- all `create_x_dataset` functions
- all `create_x_extractor` functions

## How has this been tested?

Code reorganisation so the original tests should cover it.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves #584 

## Breaking changes

Code using the `create_x_extractor` or `create_x_dataset` functions.

## Additional Notes and Examples

My code diff seems to have got a bit messed up with the moved files, not
sure what to do about that

## Future Work / Ideas

As I was working on this, I realised there are a couple more things I'd
like to change:
1) I'd like to move the `mode` and `in_memory` parameters into the data
config. This way the validation of the patching strategies can be done
by pydantic (and we can possibly have a convenience function for
converting between predict and training configs). And to me it makes
sense to have `in_memory` in the config because using the iterable tiff
loading changes the order of the patches.
2) I think it would be better to move all the initialisation of the
dataset components out of the CAREamics dataset, so we end up with
something like the example below. and the initialisation of all these
components happens in the `create_dataset` function. (Or maybe it can be
a classmethod of the dataset).

    ```python
    class CareamicsDataset(Dataset, Generic[GenericImageStack]):
        def __init__(
            self,
            data_config: NGDataConfig,
            input_extractor: PatchExtractor[GenericImageStack],
            target_extractor: PatchExtractor[GenericImageStack] | None,
            mask_extractor: PatchExtractor[GenericImageStack] | None,
            patching_strategy: PatchingStrategy,
            normalization: Normalization | None,
            augmentations: list[Augmentations],
            patch_filter: PatchFilterProtocol | None,
            coord_filter: CoordFilterProtocol | None,
        ) -> None: ...
    ```



---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>