Chore: update tests to prevent writing files outside of tmp folders (#628)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

Following https://github.com/CAREamics/careamics/issues/601, I
identified a few tests that create a `checkpoint` and `csv_logger`
folders in the working directory rather than in a `tmp_path` during
testing. I updated the tests to prevent that.

Closes https://github.com/CAREamics/careamics/issues/601