Chore(config): Add imports to `__init__` (#633)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

The config models can be deeply buried, this PR brings about more
reasonable import levels in the config modules.

I have not updated the imports throughout the code, this can be done
progressively. This was motivated by fixing imports in the example CI.