Feat(Next-Gen Dataset): return sources from `convert_prediction` in order (#631)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: 
> - Add `return_sources` to `convert_prediction` to get a list of the
corresponding sources.
> - Ensure that `stitch_prediction` and `combine_samples` return
`data_idx` in order.

Although we expect the sources and tiles to be passed and returned in
the same order, tiles may be shifted. We currently cover this
possibility, to in addition ensure that users can track which stitched
prediction corresponds to which source (in case multiple worker screw up
the order of the sources), this PR simply allows returning a list of
source from `convert_prediction`.

## Why this would matter?

With the following snippet, except believing that the stitched
predictions are returned in the exact order the sources where given to
the `inf_data_module`, there is no way to track back what
`convert_prediction` will return.

```python
tiled_predictions = trainer.predict(model, datamodule=inf_data_module)
stitched_predictions = convert_prediction(
     tiled_predictions,
     tiled=True,
)
```

Indeed if the order of `data_idx` is not sorted, then
`stitch_prediction` will sort it and return arrays in the sorted order.