Feat(Next-Gen Dataset): allow selecting which channels to extract (#640)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Allow specifying channels from which to extract data, rather
than always pulling all of them.

This PR introduces the `channels: Sequence[int] | None` parameter in the
`NGDataConfig`, which is propagated from the `CAREamicsDataset` all the
way to `ImageStack.extract_channel_patch`. Users can specify a single
channel, a list of channels, or `None` to select all of them.

Previously, `channel_idx` had been introduced and only allowed a single
channel or all, and was not accessible.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

By user request, this can be a convenient feature to not have to re-save
data in order to avoid training/predicting from a particular channel.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

`channels` parameter can now be found throughout the configuration,
`CAREamicsDataset`, `PatchExtractor`, patch extractor constructor, and
`ImageStack`.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

- new `channels` parameter in `NGDataConfig`
- `CAREamicsDataset` calls `patch_extractor.extract_channel_patch`
- `channel_idx` has been renamed to `channels` and is now `Sequence[int]
| None`
- All `ImageStack` now deal with `channels` as a sequence or `None`

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- `channels` parameter in `NGDataConfig`

### Modified features or files

<!-- List important modified features or files. -->
- All `ImageStack` and patch extractors


## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Implemented tests.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves https://github.com/CAREamics/careamics/issues/636

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

The behaviour of `extract_channel_patch` is now different since the type
of `channels` (previously `channel_idx`) has changed.

## Additional Notes and Examples

<!-- Provide any additional information that will help reviewers
understand the
changes. This can be links to documentations, forum posts, past
discussions etc. -->

One consequence is that `input` and `target` now have the same number of
channels if `channels is not None`.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)