Feat(Next-Gen Dataset): add `mode` parameter to `NGDataConfig` and add config conversion (#645)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Move `mode` and patching strategy validation from
`CAREamicsDataset` to `NGDataConfig`, and add convenience functions to
convert from training to validation or prediction `NGDataConfig`.

Moving `mode` to the config allows moving validation to the Pydantic
configuration. This ensures better coherence of the configuration and of
the pipeline.

A nice idea raised in https://github.com/CAREamics/careamics/issues/639
is to add a convenience function to convert a configuration from one
mode to the other.

Here, I chose to only support `training` to `validating` or
`predicting`, as I don't see any other use case. Additionally,
`training` requires more parameters and validation, while `validating`
and `predicting` have simple parameter sets.

Notes and potential issues:
- One problem with the current design choice is that to ask for all
channels (when `channels` was not `None` previously) when converting the
configuration, one cannot specify `channels=None`, as this would just
keep the previous `channels`... Solution is probably to pass an empty
list (counter-intuitive) or to specify the channels (e.g. `channels =
[0, 1, 2]`).
- Probably the warning should simply say that errors will be raised,
rather than performance drop, as I believe (untested as of yet) that
both UNet and LVAE will throw an error.
- A consequence of the conversion in the `CAREamicsDataModule` is that
if the data does not need patching (or different batching), prediction
data could also be passed to `CAREamicsDataModule` and predicted on
directly after training. This is untested for the moment, and we would
need to add some warnings.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

Validation of the patching strategies was done in the
`CAREamicsDataset`, but it belongs to the Pydantic configuration.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

- `mode` is now a parameter in `NGDatasConfig`
- `NGDatasConfig` has a `self.convert_mode` that allows converting
training configurations into validation or prediction, and passing some
parameters to be swapped in the configuration.
- A new `FixedRandomPatchingConfig` has been implemented.
- A new factory for creating patching strategies from configuration,
`create_patching_strategy`, has also been implemented.
- `TilingStrategy` `tile_size` parameter has been renamed to
`patch_size` to be coherent with the configuration and simplify
instantiating it in the factory.

Small additional changes:
- `PatchExtractor` parameter `shape` has been renamed `shapes`.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Tests only.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves https://github.com/CAREamics/careamics/issues/639

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

Any code directly calling `NGDataset`, `CAREamicsDataset`, and
`TilingStrategy`.

## Additional Notes and Examples

<!-- Provide any additional information that will help reviewers
understand the
changes. This can be links to documentations, forum posts, past
discussions etc. -->

Remaining TODOs:
- [ ] Assess warnings "degraded performance", it may just be an error.
- [ ] Test whether a single `CAREamicsDataModule` can be used


---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: Melisande Croft <63270704+melisande-c@users.noreply.github.com>