Feat(Next-Gen Dataset): Better Zarr tile writing (prediction with `S`, sharding/chunking, error handling) (#653)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: This PR fixed some important over sights with the Zarr tile
writing, and adds or improves sharding, chunking and error handling.

Sharding is useful to reduce the number of files being written to a zarr
folder. Currently, the Zarr tile write ignores (literally) the Zarr
source shard size. Furthermore, several problems were identified:

- `chunks` and `shards` dimensions order can (in theory) be different
from the `data_shape` received by the prediction writer, whether it is
because of dimensions order not being natively `SCZYX`, because of
additional axes (e.g. `T`), or missing customary axes (e.g. `C` or `S`).
e.g. `STYX` in the original data, `SYX` in tiles received from
prediction).
- `chunks` and `shards` may not exist in the source, e.g. `array` or
`tiff` source.
- Adding `chunks` as a parameter to all `ImageStack` and
`ImageRegionData` is violating the spirit of these classes by forcing
unrelated parameters onto some implementations. This can get out of
hands and lead to technical debt.
- The tile writing is not compatible with `S` dimension due to a faulty
implementation and lack of test coverage.

What this PR does is:

- Make tile writing compatible with `S` dimension.
- Remove random `squeeze()` during prediction writing and explicitly
remove only the axes that are not present in `axes` (zarr tile writer).
- Implement in `ZarrImageStack` the reshaping of the `chunks` in order
to be compatible with the `data_shape` and prediction data dimension
order. This also takes care of the putative presence of both `S` and `T`
in chunking.
- Remove `chunks` for `ImageStack` implementations (except Zarr) and
`ImageRegionData`
- Add `additional_metadata` in `ImageRegionData` to propagate `shards`
and `chunks` for `ZarrImageStack`.
- Add `shards` and `chunks` handling in `ZarrImageStack`, including when
these are `None` (see `_auto_chunking`).
- Raise an error when the source is not a URI.

### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

Fixing some problems (`S` dimension prediction, error handling when
source is not zarr, mismatch between chunks and data shape, chunks in
the presence of T), and adding new features (sharding).

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

- `chunks` and `shards` are passed though `additional_metadata` in
`ImageRegionData`
- `ZarrImageStack` reshapes `shards` and `chunks` the same way as it
does `data_shape`
- `WriteTilesZarr` writes `shards`, and can auto chunk if `chunks` are
None (will be useful when implementing prediction from different
sources, see Notes), and is now compatible with `S` dims.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

New tests to image stack and zarr tile writing.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves https://github.com/CAREamics/careamics/issues/650

## Notes

- For https://github.com/CAREamics/careamics/issues/654, the reshaping
of `shards` and `chunks` might need to be reverted. However, the
argument to implement it here is two-fold: prevent potential problems in
the mean time (mismatching dimension orders), takes care of the `T`
dimension missing.
- Auto-chunking is not currently used because it is only triggered when
`chunks` is `None`, meaning when the source is not Zarr. The full
compatibility requires `source: str | Path | Literal["array"]` to be
handled properly, which will be a subsequent PR:
https://github.com/CAREamics/careamics/issues/648.
- In order to propagate other type of metadata
(https://github.com/CAREamics/careamics/issues/652),
`additional_metadata` in `ImageRegionData` may be used.
- Following this PR, the case where `C` is in `axes` but only a single
channel is extracted will cause `data_shape` when writing to Zarr to
have a singleton channel. However, when writing to the array, this may
be an issue. Currently, I chose to ignore this problem as it should be
tackled in https://github.com/CAREamics/careamics/pull/651.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

[UPDATE 09.12]: updated the description to fit the current state of the
PR.